Refactor Send logic into SendManager
